move ResponseWaiter into its own source file
